Enhances performance of -a option by printing as soon as each PEL 
is parsed. 
This change minimizes the overhead associated with storing and 
processing large strings, resulting in faster output generation.

Tested Non-BMC env:
```bash
time python3 -m pel.peltool.peltool -a -p 3kpels/
1m3.707s    ->  0m38.749s
```

Optimize performance by implementing caching mechanisms for
component IDs, user data parsers, callout parsers, and SRC
modules.
These changes reduce redundant module imports by storing previously
loaded modules in dictionaries, streamlining the parsing process
and enhancing overall efficiency.

Tested:
```bash
Non-BMC env:
-n: 0.941s    ->  0.283s
-l: 6.100s    ->  4.225s
-a: 52.740s   ->  17.371s

BMC env:
-n: 4.287s   ->  4.217s
-l: 1m24.911s ->  14.125s
-a: 1m37.431s ->  24.958s
```

Signed-off-by: Harsh Agarwal Harsh.Agarwal@ibm.com