@spinler 2024-10-08
When I used the -a and redirected it to a file, when it tried to parse that file the error showed up within the JSON output:

```
            "00000650     46737043  6B73756D  00000000  A13DDA96     FspCksum.....=..",
            "00000660     4F757243  6B73756D  00000000  00000000     OurCksum........"
        ]
    }
}Failed to parse Private Header, section ID = 5b0a

]
```

And also with a garbage file named 1_1 in that dir it shows up in the middle:
```
            "Jan 04 19:40:31 systemd[1]: phosphor-regulators.service: Sent signal SIGHUP to main process 483 (phosphor-regula) on client request."
        ]
    }
}Exception: No PEL parsed for 1_1: range check failure
,
{
    "Private Header": {
```

I think probably printing those to stderr would at least get them out of the JSON?