@spinler 2024-10-09
Also here I think the commit message can just say what the fix is, and then in the body it can say it's for performance reasons.