@harsh-agarwal1 2024-10-16
Addressed [comment](https://github.com/ibm-openbmc/openpower-pel-parsers/pull/69#issuecomment-2399975352) in #70 .